1. Moving the predicate handler back to Getter method\n2. Updates to use the Generic methods
