+++ src/main/java/com/jbjohn/model/Getter.java
@@ -28,7 +28,7 @@ public class Getter {
                 return request;
             }
             if (!predicate.equals("")) {
-                return Generic.processPredicate(request, predicate);
+                return processPredicate(request, predicate);
             }
             if (key.matches("^-?\\d+$")) {
                 int index = Integer.parseInt(key);
@@ -60,20 +60,84 @@ public class Getter {
 
         List<String> stringList = Generic.getKeyList(path);
         if (stringList.size() > 1) {
-            String key = stringList.get(0);
-            if (key.startsWith("[") && key.endsWith("]")) {
-                key = key.replace("[", "").replace("]", "");
-            }
+            String key = Generic.trimKey(stringList.get(0));
             map = search(map, key);
-            String newKey = Generic.newKey(stringList);
-            path = newKey;
+            path = Generic.newKey(stringList);
         } else {
-            if (path.startsWith("[") && path.endsWith("]")) {
-                path = path.replace("[", "").replace("]", "");
-            }
+            path = Generic.trimPath(path);
             return search(map, path);
         }
         return searchByPath(map, path);
     }
 
+    public static Object processPredicate(Object map, String predicate) {
+
+        int operation = 0;
+        String predicateKey = "";
+        String predicateValue = "";
+
+        if (predicate.contains("==")) {
+            operation = 1;
+            List<String> predicateList = Arrays.asList(predicate.split("=="));
+            predicateKey = predicateList.get(0);
+            predicateValue = predicateList.get(1);
+        } else if (predicate.contains(">")) {
+            operation = 2;
+            List<String> predicateList = Arrays.asList(predicate.split(">"));
+            predicateKey = predicateList.get(0);
+            predicateValue = predicateList.get(1);
+        } else if (predicate.contains("<")) {
+            operation = 3;
+            List<String> predicateList = Arrays.asList(predicate.split("<"));
+            predicateKey = predicateList.get(0);
+            predicateValue = predicateList.get(1);
+        } else if (predicate.contains("=")) {
+            operation = 1;
+            List<String> predicateList = Arrays.asList(predicate.split("="));
+            predicateKey = predicateList.get(0);
+            predicateValue = predicateList.get(1);
+        }
+
+        if (map instanceof ArrayList) {
+            ArrayList request = (ArrayList) map;
+            ArrayList response = new ArrayList();
+
+            for (Object item : request) {
+                Map<String, Object> itemMap = (Map<String, Object>) item;
+                if (itemMap.containsKey(predicateKey)) {
+
+                    int valueInt = 0;
+                    int predicateInt = 0;
+                    if (((String) itemMap.get(predicateKey)).matches("^-?\\d+$") && predicateValue.matches("^-?\\d+$")) {
+                        valueInt = Integer.parseInt((String) itemMap.get(predicateKey));
+                        predicateInt = Integer.parseInt(predicateValue);
+                    }
+
+                    switch (operation) {
+                        case 1:
+                            if (itemMap.get(predicateKey).equals(predicateValue)) {
+                                response.add(itemMap);
+                            }
+                            break;
+                        case 2:
+                            if (valueInt > predicateInt) {
+                                response.add(itemMap);
+                            }
+                            break;
+                        case 3:
+                            if (valueInt < predicateInt) {
+                                response.add(itemMap);
+                            }
+                            break;
+                        default:
+                            break;
+                    }
+                }
+            }
+
+            return response;
+        }
+
+        return map;
+    }
 }
